# Conflicts:
#	WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/include/vector_tiles/SLDSymbolizers.h
#	WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/vector_tiles/SLDStyleSet.m

+++ WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/MaplyQuadTracker.mm
@@ -92,7 +92,7 @@ public:
 
 - (int)numTiles
 {
-    return tiles.size();
+    return (int)tiles.size();
 }
 
 - (void)tiles:(MaplyQuadTrackerPointReturn *)tilesInfo forPoints:(int)numPts
